test(core-api): extract reminder integration harness
