+++ apps/core-api/test/core.integration.test.ts
@@ -1,192 +1,16 @@
-import { beforeAll, afterAll, describe, expect, test, vi } from 'vitest';
-import { INestApplication, ValidationPipe } from '@nestjs/common';
-import { Test } from '@nestjs/testing';
+import { describe, expect, test, vi } from 'vitest';
 import request from 'supertest';
 import { createServer } from 'node:http';
 import { createHmac } from 'node:crypto';
-import { AppModule } from '../src/app.module';
-import { PrismaService } from '../src/prisma/prisma.service';
 import { AuthService } from '../src/auth/auth.service';
-import { ReminderDeliveryService } from '../src/tasks/reminder-delivery.service';
-import { TaskRetentionService } from '../src/tasks/task-retention.service';
-import { WebhookDeliveryService } from '../src/webhooks/webhook-delivery.service';
-import { CorrelationIdMiddleware } from '../src/common/correlation.middleware';
-import { RecurringTaskWorker } from '../src/recurring-tasks/recurring-task.worker';
+import { setupCoreIntegrationSuite } from './core-integration/testkit';
+import { registerBootstrapAndReminderIntegrationTests } from './core-integration/bootstrap-and-reminders';
 
 describe('Core API Integration', () => {
-  let app: INestApplication;
-  let prisma: PrismaService;
-  let token: string;
-  let reminderWorker: ReminderDeliveryService;
-  let retentionWorker: TaskRetentionService;
-  let webhookWorker: WebhookDeliveryService;
-  let recurringWorker: RecurringTaskWorker;
+  const { app, prisma, token, reminderWorker, retentionWorker, webhookWorker, recurringWorker } =
+    setupCoreIntegrationSuite();
 
-  beforeAll(async () => {
-    process.env.NODE_ENV = 'test';
-    process.env.DEV_AUTH_ENABLED = 'true';
-    process.env.DEV_AUTH_SECRET = 'atlaspm-integration-secret-123';
-    process.env.COLLAB_JWT_SECRET = 'collab-jwt-secret';
-    process.env.COLLAB_SERVICE_TOKEN = 'collab-service-secret';
-    process.env.COLLAB_SERVER_URL = 'ws://localhost:18080';
-    process.env.SEARCH_ENABLED = 'false';
-    process.env.REMINDER_WORKER_ENABLED = 'false';
-    process.env.TASK_RETENTION_WORKER_ENABLED = 'false';
-    process.env.TASK_RETENTION_DAYS = '30';
-    process.env.WEBHOOK_DELIVERY_WORKER_ENABLED = 'false';
-    process.env.WEBHOOK_DELIVERY_BASE_DELAY_MS = '0';
-    process.env.WEBHOOK_DELIVERY_MAX_DELAY_MS = '0';
-    process.env.WEBHOOK_DELIVERY_MAX_ATTEMPTS = '2';
-    process.env.WEBHOOK_SIGNING_SECRET = 'webhook-test-secret';
-    process.env.RECURRING_WORKER_ENABLED = 'false';
-    process.env.DATABASE_URL =
-      process.env.DATABASE_URL ?? 'postgresql://atlaspm:atlaspm@localhost:55432/atlaspm?schema=public';
-
-    const moduleRef = await Test.createTestingModule({ imports: [AppModule] }).compile();
-    app = moduleRef.createNestApplication();
-    app.use(new CorrelationIdMiddleware().use);
-    app.useGlobalPipes(
-      new ValidationPipe({ whitelist: true, transform: true, forbidNonWhitelisted: true }),
-    );
-    await app.init();
-    prisma = moduleRef.get(PrismaService);
-    reminderWorker = moduleRef.get(ReminderDeliveryService);
-    retentionWorker = moduleRef.get(TaskRetentionService);
-    webhookWorker = moduleRef.get(WebhookDeliveryService);
-    recurringWorker = moduleRef.get(RecurringTaskWorker);
-
-    await prisma.$connect();
-    const auth = moduleRef.get(AuthService);
-    token = await auth.mintDevToken('test-user', 'test@example.com', 'Test User');
-  });
-
-  afterAll(async () => {
-    await app.close();
-  });
-
-  test('concurrent first authenticated requests bootstrap a fresh user without 500s', async () => {
-    const auth = app.get(AuthService);
-    const userId = `concurrent-user-${Date.now()}`;
-    const freshToken = await auth.mintDevToken(
-      userId,
-      `${userId}@example.com`,
-      'Concurrent User',
-    );
-
-    const responses = await Promise.all(
-      Array.from({ length: 8 }, (_, index) =>
-        request(app.getHttpServer())
-          .get(index % 2 === 0 ? '/workspaces' : '/me')
-          .set('Authorization', `Bearer ${freshToken}`),
-      ),
-    );
-
-    for (const response of responses) {
-      expect(response.status).toBe(200);
-    }
-
-    const memberships = await prisma.workspaceMembership.findMany({
-      where: { userId },
-    });
-    expect(memberships).toHaveLength(1);
-  });
-
-  test('users can configure reminder delivery preferences and opt out of reminder delivery', async () => {
-    const auth = app.get(AuthService);
-    const reminderUserId = `reminder-pref-user-${Date.now()}`;
-    const reminderToken = await auth.mintDevToken(
-      reminderUserId,
-      `${reminderUserId}@example.com`,
-      'Reminder Preference User',
-    );
-
-    await request(app.getHttpServer()).get('/me').set('Authorization', `Bearer ${reminderToken}`).expect(200);
-
-    const workspaceRes = await request(app.getHttpServer())
-      .get('/workspaces')
-      .set('Authorization', `Bearer ${reminderToken}`)
-      .expect(200);
-    const workspaceId = workspaceRes.body[0].id as string;
-
-    const projectRes = await request(app.getHttpServer())
-      .post('/projects')
-      .set('Authorization', `Bearer ${reminderToken}`)
-      .send({ workspaceId, name: 'Reminder Preferences Project' })
-      .expect(201);
-    const projectId = projectRes.body.id as string;
-
-    const sectionsRes = await request(app.getHttpServer())
-      .get(`/projects/${projectId}/sections`)
-      .set('Authorization', `Bearer ${reminderToken}`)
-      .expect(200);
-    const defaultSectionId =
-      (sectionsRes.body.find((section: { isDefault: boolean }) => section.isDefault) ??
-        sectionsRes.body[0])?.id as string;
-
-    const taskRes = await request(app.getHttpServer())
-      .post(`/projects/${projectId}/tasks`)
-      .set('Authorization', `Bearer ${reminderToken}`)
-      .send({
-        sectionId: defaultSectionId,
-        title: 'Reminder preference delivery test',
-        dueAt: new Date(Date.now() + 86_400_000).toISOString(),
-      })
-      .expect(201);
-    const taskId = taskRes.body.id as string;
-
-    const initialPreferences = await request(app.getHttpServer())
-      .get('/me/reminder-preferences')
-      .set('Authorization', `Bearer ${reminderToken}`)
-      .expect(200);
-    expect(initialPreferences.body).toEqual({
-      enabled: true,
-      defaultLeadTimeMinutes: 60,
-    });
-
-    const updatedPreferences = await request(app.getHttpServer())
-      .put('/me/reminder-preferences')
-      .set('Authorization', `Bearer ${reminderToken}`)
-      .send({ enabled: false, defaultLeadTimeMinutes: 1440 })
-      .expect(200);
-    expect(updatedPreferences.body).toEqual({
-      enabled: false,
-      defaultLeadTimeMinutes: 1440,
-    });
-
-    await request(app.getHttpServer())
-      .put(`/tasks/${taskId}/reminder`)
-      .set('Authorization', `Bearer ${reminderToken}`)
-      .send({ remindAt: new Date(Date.now() - 60_000).toISOString() })
-      .expect(200);
-
-    const disabledDeliveryCount = await reminderWorker.processDueReminders(new Date());
-    expect(disabledDeliveryCount).toBe(0);
-
-    const reminderWhileDisabled = await request(app.getHttpServer())
-      .get(`/tasks/${taskId}/reminder`)
-      .set('Authorization', `Bearer ${reminderToken}`)
-      .expect(200);
-    expect(reminderWhileDisabled.body.sentAt).toBeNull();
-
-    const reenabledPreferences = await request(app.getHttpServer())
-      .put('/me/reminder-preferences')
-      .set('Authorization', `Bearer ${reminderToken}`)
-      .send({ enabled: true, defaultLeadTimeMinutes: 15 })
-      .expect(200);
-    expect(reenabledPreferences.body).toEqual({
-      enabled: true,
-      defaultLeadTimeMinutes: 15,
-    });
-
-    const deliveredCount = await reminderWorker.processDueReminders(new Date());
-    expect(deliveredCount).toBe(1);
-
-    const reminderAfterReenable = await request(app.getHttpServer())
-      .get(`/tasks/${taskId}/reminder`)
-      .set('Authorization', `Bearer ${reminderToken}`)
-      .expect(200);
-    expect(Boolean(reminderAfterReenable.body?.sentAt)).toBe(true);
-  });
+  registerBootstrapAndReminderIntegrationTests({ app, prisma, token, reminderWorker, retentionWorker, webhookWorker, recurringWorker });
 
   test('project/member/sections/tasks/rules/reorder/audit/outbox flow', async () => {
     await request(app.getHttpServer()).get('/me').set('Authorization', `Bearer ${token}`).expect(200);
